Task_2: Added eslint rules for imports, fixed comments, small refactoring
