+++ my-app/angular.json
@@ -13,6 +13,9 @@
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
           "options": {
+            "allowedCommonJsDependencies": [
+              "lodash"
+            ],
             "outputPath": "dist/my-app",
             "index": "src/index.html",
             "main": "src/main.ts",
