+++ my-app/src/app/components/courses/courses.component.ts
@@ -1,10 +1,12 @@
-import {Component, Input, Output, EventEmitter} from '@angular/core';
-import {Course} from "../../app.component";
+import {Component, Input, Output, EventEmitter, ChangeDetectionStrategy} from '@angular/core';
+
+import {Course} from "../../shared";
 
 @Component({
   selector: 'app-courses',
   templateUrl: './courses.component.html',
-  styleUrls: ['./courses.component.scss']
+  styleUrls: ['./courses.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class CoursesComponent {
   @Input() course: Course | undefined;
@@ -12,21 +14,5 @@ export class CoursesComponent {
 
   public deleteButtonClicked(value: number) {
     this.deleteCourse.emit(value)
-  }
-
-  public durationHandler(duration: number): string {
-    const MIN_IN_HOUR: number = 60;
-    const divisionOfDuration: number = duration / MIN_IN_HOUR;
-    const remainderOfDivision: number = duration % MIN_IN_HOUR;
-    let handledDuration: string;
-
-    if (duration >= MIN_IN_HOUR) {
-      remainderOfDivision
-        ? handledDuration = `${Math.floor(divisionOfDuration)} h ${remainderOfDivision} min`
-        : handledDuration = `${divisionOfDuration} h`
-    } else {
-      handledDuration = `${duration} min`
-    }
-    return handledDuration
   }
 }
