+++ my-app/src/app/components/load-more-button/load-more-button.component.html
@@ -1,1 +1,1 @@
-<button class="load-more-button" (click)="loadMoreButtonClicked()">LOAD MORE</button>
+<button (click)="loadMoreButtonClicked()">LOAD MORE</button>
