+++ my-app/src/app/components/load-more-button/load-more-button.component.scss
@@ -1,4 +1,4 @@
-.load-more-button {
+button {
   width: 100%;
   background-color: white;
   border-radius: 20px;
