+++ my-app/src/app/components/load-more-button/load-more-button.component.ts
@@ -1,13 +1,15 @@
-import { Component } from '@angular/core';
+import {ChangeDetectionStrategy, Component} from '@angular/core';
 
 @Component({
   selector: 'app-load-more-button',
   templateUrl: './load-more-button.component.html',
-  styleUrls: ['./load-more-button.component.scss']
+  styleUrls: ['./load-more-button.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class LoadMoreButtonComponent {
 
-  constructor() {}
+  constructor() {
+  }
 
   public loadMoreButtonClicked(): void {
     console.log("Load More Btn was clicked")
