+++ my-app/src/app/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './pipes/duration-handler.pipe';
+export * from './types';
