+++ my-app/src/app/shared/types.ts
@@ -1,0 +1,13 @@
+export interface Course {
+  id: number;
+  title: string;
+  creationDate: Date;
+  duration: number;
+  description: string;
+}
+
+export interface User {
+  id: number;
+  firstName: string;
+  lastName: string;
+}
